icecc.bbclass: add a class to support IceCC compile farms
This class is somewhat based on what is present in org.oe.dev, but is
largely rewritten. Toolchains packages are built in runtime (and not
while parsing), icecc-create-env is built automagically, etc.

The only problem with current oe-core is that this class is incompatible
with apt-native, so it should be BBMASK'ed in your local.conf

To use this class put the following lines in your conf/local.conf:
ICECC_PARALLEL_MAKE = "-j 15"
INHERIT += "icecc"
BBMASK="apt-native"

Signed-off-by: Dmitry Eremin-Solenikov <dbaryshkov@gmail.com>